🔧 Refactor bottom sheet layouts with shared component
- Create LinkActionLayoutView to centralize bottom sheet layout logic
- Replace duplicate layout configurations in link-action and link-input routes
- Simplify route layout components by using shared layout view
