+++ app/bottom-sheet/_layout.tsx
@@ -1,20 +1,8 @@
 import { Stack } from "expo-router";
 
-export default function BottomSheetLayout() {
-  const sheetScreenOptions = {
-    presentation: "formSheet" as const,
-    contentStyle: {
-      backgroundColor: "transparent",
-      height: 280,
-    },
-    sheetGrabberVisible: false,
-    gestureDirection: "vertical" as const,
-    animation: "slide_from_bottom" as const,
-    headerShown: false,
-    sheetInitialDetentIndex: 0,
-    sheetAllowedDetents: [0.38, 0.38],
-  };
+import { sheetScreenOptions } from "@/components/layout/bottom-sheet/constants/screenOption";
 
+export default function BottomSheetLayout() {
   return (
     <Stack>
       <Stack.Screen name="link-input" options={sheetScreenOptions} />
